# docs/api/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fluvia API",
    "version": "1.0.0-sandbox",
    "description": "Contrato PARCIAL y versionado del plano de integración (F3-02): payment intents. Se amplía por incremento; un test de contrato verifica que cada (path, method) declarado existe en el servidor y exige autenticación. Errores: SIEMPRE el sobre {error:{type,code,message,details?,request_id}} del catálogo v1 (api-errors.md). Mutaciones: header Idempotency-Key OBLIGATORIO (contrato idempotency.md: replay exacto con el mismo par key+payload; 422 idempotency_key_reuse si el payload difiere; 409 processing_in_flight si sigue en vuelo)."
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key fluvia_sk_test_* (scopes: read, payments:write)"
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 255
        }
      }
    },
    "schemas": {
      "PaymentIntent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "object": {
            "const": "payment_intent"
          },
          "merchant_id": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "integer",
            "description": "Unidades MENORES de la moneda"
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$"
          },
          "status": {
            "enum": [
              "created",
              "requires_payment_method",
              "requires_confirmation",
              "requires_action",
              "processing",
              "authorized",
              "partially_captured",
              "succeeded",
              "failed",
              "canceled",
              "partially_refunded",
              "refunded"
            ]
          },
          "capture_method": {
            "enum": ["automatic", "manual"]
          },
          "amount_captured": {
            "type": "integer"
          },
          "amount_refunded": {
            "type": "integer"
          },
          "failure_code": {
            "type": ["string", "null"]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string"
              },
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {},
              "request_id": {
                "type": "string"
              }
            },
            "required": ["type", "code", "message", "request_id"]
          }
        },
        "required": ["error"]
      }
    }
  },
  "security": [
    {
      "apiKey": []
    }
  ],
  "paths": {
    "/v1/payment_intents": {
      "post": {
        "summary": "Crear payment intent (scope payments:write; idempotente)",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["merchant_id", "amount", "currency"],
                "properties": {
                  "merchant_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "amount": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "currency": {
                    "type": "string",
                    "pattern": "^[A-Z]{3}$"
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "capture_method": {
                    "enum": ["automatic", "manual"]
                  },
                  "metadata": {
                    "type": "object",
                    "maxProperties": 20,
                    "additionalProperties": {
                      "type": "string",
                      "maxLength": 500
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Intent creado (o replay exacto)"
          },
          "400": {
            "description": "validation_error | idempotency_key_required"
          },
          "404": {
            "description": "not_found (merchant ajeno o inexistente)"
          },
          "409": {
            "description": "processing_in_flight"
          },
          "422": {
            "description": "idempotency_key_reuse"
          }
        }
      },
      "get": {
        "summary": "Listar payment intents del tenant (scope read)",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista, más reciente primero"
          }
        }
      }
    },
    "/v1/payment_intents/{id}": {
      "get": {
        "summary": "Recuperar un payment intent (scope read)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "El intent"
          },
          "404": {
            "description": "not_found (ajeno e inexistente indistinguibles)"
          }
        }
      }
    },
    "/v1/payment_intents/{id}/cancel": {
      "post": {
        "summary": "Cancelar un payment intent (scope payments:write; idempotente). La FSM decide: sólo estados con transición legal a canceled.",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Intent cancelado (o replay exacto)"
          },
          "404": {
            "description": "not_found"
          },
          "409": {
            "description": "invalid_state_transition (estado terminal u otro sin transición legal)"
          }
        }
      }
    },
    "/v1/payment_intents/{id}/confirm": {
      "post": {
        "summary": "Confirmar un payment intent (scope payments:write; idempotente). ASINCRONO por contrato: responde el intent en processing con attempt_id; el estado final se lee via GET. tokens del sandbox: tok_approve, tok_decline, tok_decline_insufficient, tok_timeout (indeterminado).",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["payment_method_token"],
                "properties": {
                  "payment_method_token": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Intent en processing + attempt_id (o replay exacto)"
          },
          "404": {
            "description": "not_found"
          },
          "409": {
            "description": "invalid_state_transition (ya processing, terminal, etc.)"
          },
          "422": {
            "description": "idempotency_key_reuse"
          }
        }
      }
    },
    "/v1/webhook_endpoints": {
      "post": {
        "summary": "Crear endpoint de webhook (scope webhooks:manage). El secreto whsec_ se muestra UNA sola vez. URL: https obligatorio fuera de local; guard SSRF completo en cada entrega.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["url"],
                "properties": {
                  "url": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "events": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Topics del catálogo (webhook-delivery.md §5); vacío = todos"
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 200
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Endpoint creado (incluye secret una única vez)"
          },
          "400": {
            "description": "validation_error (URL insegura o topic desconocido)"
          }
        }
      },
      "get": {
        "summary": "Listar endpoints (scope webhooks:manage); jamás incluye secretos",
        "responses": {
          "200": {
            "description": "Lista"
          }
        }
      }
    },
    "/v1/webhook_endpoints/{id}/rotate": {
      "post": {
        "summary": "Rotar el secreto (scope webhooks:manage). El anterior sigue firmando durante la ventana de gracia (24 h): cada entrega lleva v1=nuevo,v1=viejo.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Nuevo secret (una única vez)"
          },
          "404": {
            "description": "not_found"
          }
        }
      }
    },
    "/v1/webhook_endpoints/{id}/disable": {
      "post": {
        "summary": "Deshabilitar endpoint (scope webhooks:manage); su cola pendiente muere sin intentos de red",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Endpoint deshabilitado"
          },
          "404": {
            "description": "not_found"
          }
        }
      }
    }
  }
}